keycloak load changed
